Move out of logs folder
