+++ test/entity/resources/config_metrics_resource.json
